refactore: rebuild notFound component to Message to show different messages
