+++ src/components/Message/Message.module.css
@@ -1,4 +1,4 @@
-.not-found {
+.message {
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -7,7 +7,7 @@
   margin: 0 auto;
 }
 
-.not-found-title {
+.message-title {
   margin: 0;
   margin-bottom: 20px;
   font-size: 48px;
@@ -15,7 +15,7 @@
   color: #fff;
 }
 
-.not-found-text {
+.message-text {
   max-width: 540px;
   margin: 0;
   font-size: 20px;
