+++ src/main.tsx
@@ -2,10 +2,10 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import Layout from './layouts/Layout';
+import Message from './components/Message/Message';
 import MainPage from "./pages/mainPage/MainPage";
 import LoginPage from "./pages/loginPage/LoginPage";
 import FavoritesPage from "./pages/favoritesPage/FavoritesPage";
-import NotFound from './components/notFound/NotFound';
 import MoviePage from './pages/moviePage/MoviePage';
 import './index.css';
 
@@ -23,9 +23,9 @@ const router = createBrowserRouter([
       {
         path: '/movie/:id',
         element: <MoviePage />,
-        errorElement: <>Something went wrong, sorry.....</>,
+        errorElement: <Message type='error' />,
         loader: async ({ params }) => {
-          console.log(typeof params.id);
+          // console.log(typeof params.id);
           const { data } = await fetch(`${PREFIX}?tt=${params.id}`) as any;
           console.log(data);
           return data;
@@ -43,7 +43,7 @@ const router = createBrowserRouter([
   },
   {
     path: '*',
-    element: <NotFound />
+    element: <Message type={'error'} />
   }
 ]);
